Allocate questions algorithm implemented
Think of each users allocated questions as a tower. Each user has their
own tower. When we have a new question to allocate to "C" people, we
"sprinkle" it over the smallest towers, so that each tower increases
roughly in parallel, but still randomly. This ensures that each question
is allocated enough times, and also that they are distributed randomly.